requirements.txt
asgiref==3.5.2
boto3==1.24.61
botocore==1.27.61
Django==4.1
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.13.1
jmespath==1.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
